requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
yt-dlp>=2023.12.0
openai-whisper==20231117
lyricsgenius>=3.0.1
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.1
requests>=2.31.0
python-dotenv>=1.0.0
